call DrawableFont with font name instead of family
When using Magick as the backend, label_width() calls img.font(vFont) to get metrics and calculate text width; but the actual drawing is done with DrawableFont(name, AnyStyle, 0, AnyStretch) which takes a font family name. Thus, specifying a --font argument would either fail on the img.font() call, or DrawableFont() falls back to a default font.

PS output is not affected by the problem or the solution.

This patch means that `--font arial` may not work, but `--font arial-italic` would.

See discussions on https://aur.archlinux.org/packages/drawtiming and github tracker.
Fix #2 on GH